start of restructure of the iteration scheme.
